requirements.txt
duckdb>=0.10.0
fastapi==0.112.2
gunicorn>=21.2
httpx>=0.27.0
pydantic>=2.7
pydantic-settings>=2.2
python-jose[cryptography]>=3.3.0
slowapi>=0.1.7
tiktoken>=0.6.0
uvicorn[standard]>=0.24.0

